fix data location
