+++ front_end/src/app/(main)/labor-hub/helpers/labor_hub_posts.ts
@@ -4,53 +4,9 @@ import { cache } from "react";
 import ServerPostsApi from "@/services/api/posts/posts.server";
 import { PostWithForecasts } from "@/types/post";
 
-import { JOBS_DATA } from "../data";
+import { JOBS_DATA, LABOR_HUB_POST_IDS } from "../data";
 
 export const LABOR_HUB_REVALIDATE_SECONDS = 60 * 15;
-
-// Non-job Labor Hub post IDs. Job post IDs live in JOBS_DATA.
-export const LABOR_HUB_POST_IDS = {
-  overallEmployment: 41307,
-  activityMonitor: 41307,
-  wages: {
-    overall: 42216,
-    softwareDevelopers: 43106,
-    nurses: 43107,
-    salesRepresentatives: 43108,
-    financialSpecialists: 43109,
-    designers: 43110,
-    averageWeeklyHours: 41574,
-    aiDailyUse: 42215,
-  },
-  familyResources: {
-    twentiethPercentile: 42944,
-    fiftiethPercentile: 43042,
-    eightiethPercentile: 43043,
-  },
-  graduates: {
-    unemploymentRate: 42212,
-    underemploymentRate: 42213,
-    overallFourYearDegrees: 42220,
-    stemFourYearDegrees: 42852,
-    humanitiesFourYearDegrees: 42853,
-    tradeSchoolAndCommunityCollege: 42856,
-  },
-  economy: {
-    laborProductivity: 43087,
-    fortune500SmallWorkforce: 41324,
-    longTermUnemploymentRate: 41313,
-    laborShareOfNationalIncome: 41578,
-  },
-  research: {
-    occupationalMixChange: 42850,
-  },
-  state: {
-    washingtonOverallEmployment: 43081,
-    washingtonAerospaceSector: 43084,
-    washingtonTechnologySector: 43085,
-    washingtonHealthcareSector: 43086,
-  },
-} as const;
 
 function collectPostIds(value: unknown, ids: number[] = []): number[] {
   if (typeof value === "number") {
