+++ front_end/src/app/(main)/labor-hub/page.tsx
@@ -17,7 +17,7 @@ import {
   SectionHeader,
   ContentParagraph,
 } from "./components/section";
-import { LABOR_HUB_POST_IDS } from "./helpers/labor_hub_posts";
+import { LABOR_HUB_POST_IDS } from "./data";
 import { ActivityMonitorSection } from "./sections/activity_monitor";
 import { EngagementSection } from "./sections/engagement_section";
 import { HeroSection } from "./sections/hero";
